# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MICO360 Tasks API
  version: 1.0.0
  description: >-
    REST API for MICO360 Tasks (web app, Chrome extension, and the Phase 2 Android
    app). All responses use a JSON envelope: success is `{ "data": ... }`, errors
    are `{ "error": { "code": string, "message": string, "details"?: any } }`.
    Authenticate by sending `Authorization: Bearer <accessToken>`.
servers:
  - url: http://localhost:4000/api/v1
    description: Local development
  - url: https://task.mico360.com/api/v1
    description: Production
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /users/me/export:
    get:
      summary: Download my data — everything held about the signed-in person, as a JSON file (recorded in the audit log)
      responses:
        '200': { description: JSON file (Content-Disposition attachment) }
        '401': { description: Not signed in }
...